[REFACTOR] Moved fake communication and network tools to separate files
Fixes #92
